Updates
I have made the game look like ours again, kept the platforms from the game Kyle uploaded as barriers, added bases in the right positions, the code is currently not that tidy as there are a lot of global variables.  Also, the flag position = the player position when collected.  The player does not have a colour yet, and another thing on the to do list is to set up the text at the top of the screen for blue and red scores and the counter.